read boot-once, zero out, set bootoncelabel as default
This patch reads the boot-once environment variable from Open Firmware
and stores in a global variable, "bootoncelabel".

It then unconditionally zeros out the boot-once environment variable.

*** Note that according to 1275, setprop will create a specified
environment variable if it doesn't already exist.  A bug in nvsetenv
causes it to fail silently if you try to set an OF environment variable
that does not already exist.  However, simply booting a system running a
yaboot with this patch set compiled in will in fact create the boot-once
environment variable, which nvsetenv can later set and reset.

Signed-off-by: Dustin Kirkland <dustin.kirkland@us.ibm.com>