yaboot menu user interface
This patch allows for denoting a kernel label as a "boot-once" label
when printing the menu of labels.

The label flagged as default is marked with an asterisk (*).  If a label
matches the global bootoncelabel, I'm marking it with an ampersand (&).
I'm ambivalent about the symbol used here, or even if it is.  But I
found this bit useful when debugging and testing the functionality, and
it's trivial to add cleanly.  I chose ampersand because it's immediately
left of the asterisk on my keyboard.