use public interface to detect CAS reboots (take 2)
This is the second pass of this patch.  The fw variable name is changed
to use ibm,client-architecture-support-reboot field which is already present
in the firmware.  Please disregard the previous patch.

The firmware field used to detect CAS reboots (ibm,fw-nbr-reboots) is
really a private field that could change without warning.  A new field
ibm,#reconfig-reboots will be added as a public interface intended to
be used for this detection.  The patch will first check for
ibm,client-architecture-support-reboot if that is not found it will see
if ibm,fw-nbr-reboots is present and will use that instead.

-------