Save arguments as well as image name for CAS reboot
We were saving the image name that the user had typed (or had been
selected by default) in the "boot-last-label" property, but we were
losing any extra arguments that the user had typed after the image
name.  On a pSeries machine, if the firmware decides to do a reboot
at the client-architecture-support call, we were then rebooting to
the right image but without any extra arguments that the user typed.

This fixes the problem by saving the arguments in boot-last-label,
separated from the image name by a space.  When we use the value
of boot-last-label, we now call word_split to separate it into the
image name and arguments again.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Tony Breeds <tony@bakeyournoodle.com>